requirements.txt
alembic==1.7.5
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.8.4
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
